Improve support for connections/edges
