+++ lib/graphql/relay.rb
@@ -15,3 +15,4 @@ require 'graphql/relay/node'
 require 'graphql/relay/connection_instrumentation'
 require 'graphql/relay/connection_resolve'
 require 'graphql/relay/connection_type'
+require 'graphql/relay/type_extensions'
