+++ lib/graphql/relay/connection_type.rb
@@ -13,15 +13,17 @@ module GraphQL
       self.bidirectional_pagination = false
 
       # Create a connection which exposes edges of this type
-      def self.create_type(wrapped_type, edge_type: wrapped_type.edge_type, edge_class: GraphQL::Relay::Edge, nodes_field: ConnectionType.default_nodes_field, &block)
+      def self.create_type(wrapped_type, edge_type: nil, edge_class: GraphQL::Relay::Edge, nodes_field: ConnectionType.default_nodes_field, &block)
         custom_edge_class = edge_class
 
         # Any call that would trigger `wrapped_type.ensure_defined`
         # must be inside this lazy block, otherwise we get weird
         # cyclical dependency errors :S
         ObjectType.define do
-          name("#{wrapped_type.name}Connection")
-          description("The connection type for #{wrapped_type.name}.")
+          type_name = wrapped_type.is_a?(GraphQL::BaseType) ? wrapped_type.name : wrapped_type.graphql_name
+          edge_type ||= wrapped_type.edge_type
+          name("#{type_name}Connection")
+          description("The connection type for #{type_name}.")
           field :edges, types[edge_type], "A list of edges.", edge_class: custom_edge_class, property: :edge_nodes
 
           if nodes_field
